+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add `--quiet` argument that decreases loglevel
 * Encode hash in base64 in order to make needed URL prefix shorter at same "guessability".
   → This causes the maximum prefix length to go down from 128 to 64 (4 bit per char → 8 bit per char)
+* `list`: Add `--with-size`-switch to print file sizes.
 
 ## v0.3.1 (2020-10-03)